% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccutN_f.R
\name{ccutN_f}
\alias{ccutN_f}
\title{Correlation cut-off as function of sample size}
\usage{
ccutN_f(N)
}
\arguments{
\item{N}{is the sample size}
}
\value{
the correlation cut-off value
}
\description{
Empirical function based on simulation
}
\author{
Trey Smith, \email{treysm@umich.edu}
}
