% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmr.R
\name{cmr}
\alias{cmr}
\title{Construct Co-methylated Regions.}
\usage{
cmr(
  Mdata,
  meds = NULL,
  Iarray = c("450K", "EPIC", "EPICv2"),
  cormethod = c("pearson", "kendall", "spearman"),
  corlo = NULL,
  corhi = corlo,
  corlodst = 400,
  corhidst = 1,
  maxprbdst = 2000,
  maxdlvl = Inf,
  verbose = TRUE
)
}
\arguments{
\item{Mdata}{matrix of betas, preferably cell-type corrected, or M-values, where rows are observations and colums are array probes.}

\item{meds}{medians of probe betas, used if constraining adjacent probe maximum difference}

\item{Iarray}{which Illumina array platform, one of "450K" (default), EPIC, or "EPICv2".}

\item{cormethod}{a character string indicating which correlation is to be computed. One of "pearson" (default), "kendall", or "spearman".}

\item{corlo}{low threshold for correlation, when max distance betweek adjacent genomic CpGs is corlodst; optional - default is NULL}

\item{corhi}{high threshold for correlation, when max distance betweek adjacent genomic CpGs is corhidst; optional - default is NULL}

\item{corlodst}{the maximum distance between CpGs considered, defines low CpG denisty, default is 400bp}

\item{corhidst}{the minimum distance between CpGs considered, defines high CpG denisty, default is 1bp}

\item{maxprbdst}{Maximum bp distance between adjacent probes, default is 2000bp}

\item{maxdlvl}{the maximum difference between the medians of the adjacent probes}

\item{verbose}{whether to report progress for chromosomes, default is TRUE}
}
\value{
List of chromosome-lists of co-methylated probe set character vectors.
}
\description{
Function \code{cmr} constructs co-methylated regions based on pair-wise positive correlations above a correlation threshold.
The threshold can be optionally set to vary with genomic CpG density, in which case it increases linearly from a low value for low-density CpG background to a high value for high density background.
Pearson Corrlation is used by default, appropriate for cell-type corrected methylation betas.
}
\examples{
\dontrun{
cmr("Mdata_ProbesAreColumns_matrix")
}
}
\author{
Trey Smith, \email{treysm@umich.edu}
}
