% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmr_comp.R
\name{cmr_comp}
\alias{cmr_comp}
\title{Composite Methylation for Co-methylated Regions.}
\usage{
cmr_comp(
  cmrs,
  Mdata,
  cmethod = c("pca", "mean", "median"),
  retweights = F,
  verbose = T
)
}
\arguments{
\item{Mdata}{matrix of DNA methylation values (Beta or M), rows are observations, columns are probes}

\item{cmethod}{the aggregarion method for the composite measure. One of "pca" (default), "mean" or "median"}

\item{retweights}{whether to return the weights derived from the 1st PC, default is FALSE.  If set to TRUE, the output is a list of 2 elemnts: the composite methylation matrix, and a list of per-cmr weights}

\item{verbose}{report progress, default is TRUE}

\item{cmrs=NULL}{is the list of CMRs construced with a call to the cmr function.  If not provided, CMRs are estimated with the methylation data.}
}
\value{
matrix of CMR composite methylation values, one column per CMR.
}
\description{
Function \code{cmr_comp} estimates a composite mehtylation measure for co-methylated regions.
}
\examples{
\dontrun{
cmr_comp(CMRs,Mdata_ProbesAreColumns_matrix)
}
}
\author{
Trey Smith, \email{treysm@umich.edu}
}
