% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comeback.R
\docType{package}
\name{CoMeBackV2}
\alias{CoMeBackV2}
\alias{CoMeBackV2-package}
\title{CoMeBackV2: CoMeBack: Co-Methylation With Genomic CpG Background}
\description{
Co-methylation of adjacent array probes, taking into account genomic CpG background. Function cmr() constructs co-methylated regions based on pair-wise positive correlations above variable correlation threshold.  The threshold increases linearly from a low value for low-density CpG background to a high value for high density background.  Pearson Corrlation is used.
}
\details{
Currently, only the cmr function is provided.
}
\section{cmr}{

The \code{cmr} function constructs comethylated adjacent probe sets with correlated methylation across individuals,
with correlation threshold depending on genomic CpG density between each two probes
}

\section{cmr_comp}{

The \code{cmr_comp} function constructs a composite methylation measure for each CMR
}

\author{
\strong{Maintainer}: Trey Smith \email{treysm@umich.edu}

}
\keyword{internal}
