% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggThresh.R
\name{suggThresh}
\alias{suggThresh}
\title{Calculate suggestive P-Value threshold}
\usage{
suggThresh(Data,Columns)
}
\arguments{
\item{Data}{matrix of beta or M-Values}
\item{Columns}{list of column i.e. probe names}
}
\value{
numeric value representing a suggestive P-Value threshold
}
\description{
calculates a suggestive P-Value threshold using SVD (Singular Value Decomposition),
which estimates the number of linearly independent combinations of probes
}
\author{
Trey Smith, \email{treysm@umich.edu}
}
